Trabalhando com o controller de solicitação de compras

+++ ApiSGCOlimpiada/Models/SolicitacaoCompra.cs
@@ -7,11 +7,11 @@ namespace ApiSGCOlimpiada.Models
 {
     public class SolicitacaoCompra
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string ResponsavelEntrega { get; set; }
         public DateTime Data { get; set; }
-        public int TipoCompraId { get; set; }
-        public int EscolaId { get; set; }
+        public long TipoCompraId { get; set; }
+        public long EscolaId { get; set; }
         public TipoCompra TipoCompra { get; set; }
         public Escola Escola { get; set; }
         public string Justificativa { get; set; }
